backend/requirements.txt
Flask==3.0.2
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1
rembg
onnxruntime
flask-cors==5.0.1
werkzeug==3.1.3
gunicorn==22.0.0 